some macros to simplify if-let & let-else patterns
let-else isn't formatabble by cargo fmt and the `Some()` around the variable kind of reduces the importance of it.
we also execute some important code inside the if-let block but it doesn't look that important since it is inside an if and we continue normally otherwise.

The point is, most of these options should never be None (e.g. `connections().get(address)`), we can either panic there or just ignore
the error. Ignoring it is safer but fills the code with a lot of if-let let-else patterns.

The added macros makes the refactor somewhat similar to `?` in `ret_option()?`